# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: WASAText messaging service API for Web and Software Architecture course
  version: "1.0.0"
...
servers:
  - url: http://localhost:3000
paths:
  /session:
    post:
      tags: ["Login"]
      operationId: doLogin
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created, and an identifier is returned.
        If the user exists, the user identifier is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON body for login
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 16
                  pattern: "^[a-zA-Z0-9_]+$"
                  example: "Maria"
                  description: Username for login
      responses:
        "201":
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /conversations/{conversationId}:
    parameters:
      - $ref: "#/components/parameters/conversationId"
    get:
      tags: ["Conversations"]
      operationId: getConversation
      summary: Get conversation messages
      description: Returns all messages in a conversation, sorted in reverse chronological order. Marks messages as read for the user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User is not a member of this conversation
        "404":
          description: Conversation not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /groups:
    post:
      tags: ["Groups"]
      operationId: createGroup
      summary: Create a new group
      description: Creates a new group conversation with the creator as the first member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON body for creating a group
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
                  description: Group name
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/members:
    parameters:
      - $ref: "#/components/parameters/groupId"
    post:
      tags: ["Groups"]
      operationId: addToGroup
      summary: Add a user to the group
      description: Adds a user to the group. Only group members can add others.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON body for adding a member
              required:
                - userId
              properties:
                userId:
                  type: string
                  description: User identifier to add
                  minLength: 1
                  maxLength: 64
      responses:
        "204":
          description: User added successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Requester is not a member of this group
        "404":
          description: Group or user not found
        "500":
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/groupId"
      - $ref: "#/components/parameters/userId"
    delete:
      tags: ["Groups"]
      operationId: leaveGroup
      summary: Leave the group
      description: Removes the user from the group. Users can only remove themselves.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Left group successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User can only remove themselves
        "404":
          description: Group not found or user not a member
        "500":
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/name:
    parameters:
      - $ref: "#/components/parameters/groupId"
    put:
      tags: ["Groups"]
      operationId: setGroupName
      summary: Set group name
      description: Changes the group name. Only members can change it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON body for setting group name
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 50
      responses:
        "204":
          description: Group name changed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User is not a member of this group
        "404":
          description: Group not found
        "500":
          $ref: "#/components/responses/InternalServerError"
  /groups/{groupId}/photo:
    parameters:
      - $ref: "#/components/parameters/groupId"
    put:
      tags: ["Groups"]
      operationId: setGroupPhoto
      summary: Set group photo
      description: Sets the group photo. Only members can change it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: Binary PNG image data
          image/jpeg:
            schema:
              type: string
              format: binary
              description: Binary JPEG image data
      responses:
        "204":
          description: Group photo changed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User is not a member of this group
        "404":
          description: Group not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: User identifier returned from doLogin
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User identifier
      schema:
        type: string
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Conversation identifier
      schema:
        type: string
...
    groupId:
      name: groupId
      in: path
      required: true
      description: Group identifier
      schema:
        type: string
  schemas:
    LoginResponse:
      type: object
      description: Response object containing the login session identifier
      properties:
        identifier:
          type: string
          description: User identifier to use as Bearer token
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9]+$"
          example: "abcdef0123456789"
      required:
        - identifier
    User:
      type: object
      description: Represents a user in the system
      properties:
        id:
          type: string
          description: Unique identifier of the user
          minLength: 1
          maxLength: 64
        username:
          type: string
          description: The user's display name
          minLength: 3
          maxLength: 16
          pattern: "^[a-zA-Z0-9_]+$"
        photoUrl:
          type: string
          description: URL to the user's profile photo
          nullable: true
          maxLength: 2048
      required:
        - id
        - username
...
    MessagePreview:
      type: object
      description: A summary of a message
      properties:
        content:
          type: string
          description: The content of the message
          minLength: 1
          maxLength: 4096
        timestamp:
          type: string
          description: The time the message was sent
          format: date-time
        senderId:
          type: string
          description: ID of the user who sent the message
          minLength: 1
          maxLength: 64
      required:
        - content
        - timestamp
        - senderId
    Conversation:
      type: object
      description: Detailed conversation information including messages
      properties:
        id:
          type: string
          description: Unique identifier of the conversation
          minLength: 1
          maxLength: 64
        type:
          type: string
          description: Type of conversation
          enum: ["private", "group"]
        name:
          type: string
          description: Name of the conversation
          minLength: 1
          maxLength: 50
        photoUrl:
          type: string
          description: URL to conversation photo
          nullable: true
          maxLength: 2048
        members:
          type: array
          description: List of members in the conversation
          items:
            $ref: "#/components/schemas/User"
          minItems: 1
          maxItems: 100
        messages:
          type: array
          description: List of messages in the conversation
          items:
            $ref: "#/components/schemas/Message"
          minItems: 0
          maxItems: 1000
      required:
        - id
        - type
        - name
        - members
        - messages
    Message:
      type: object
      description: A message within a conversation
      properties:
        id:
          type: string
          description: Unique identifier of the message
          minLength: 1
          maxLength: 64
        senderId:
          type: string
          description: ID of the sender
          minLength: 1
          maxLength: 64
        senderUsername:
          type: string
          description: Username of the sender
          minLength: 3
          maxLength: 16
        type:
          type: string
          description: Type of message content
          enum: ["text", "photo"]
        content:
          type: string
          description: Text content or photo URL
          minLength: 1
          maxLength: 4096
        timestamp:
          type: string
          description: ISO 8601 timestamp
          format: date-time
        checkmarks:
          type: integer
          enum: [0, 1, 2]
          description: "0=sent, 1=received by all, 2=read by all"
        replyTo:
          allOf:
            - $ref: "#/components/schemas/MessagePreview"
            - nullable: true
        forwarded:
          type: boolean
          description: Whether the message was forwarded
        comments:
          type: array
          description: List of reactions/comments
          items:
            $ref: "#/components/schemas/Comment"
          minItems: 0
          maxItems: 100
      required:
        - id
        - senderId
        - senderUsername
        - type
        - content
        - timestamp
        - checkmarks
        - forwarded
        - comments
    Comment:
      type: object
      description: A reaction or comment on a message
      properties:
        userId:
          type: string
          description: ID of the user who commented
          minLength: 1
          maxLength: 64
        username:
          type: string
          description: Username of the commenter
          minLength: 3
          maxLength: 16
        comment:
          type: string
          description: The comment text or emoji
          minLength: 1
          maxLength: 50
      required:
        - userId
        - username
        - comment
    Group:
      type: object
      description: A group conversation
      properties:
        id:
          type: string
          description: Unique identifier of the group
          minLength: 1
          maxLength: 64
        name:
          type: string
          description: Name of the group
          minLength: 1
          maxLength: 50
        photoUrl:
          type: string
          description: URL to the group photo
          nullable: true
          maxLength: 2048
        members:
          type: array
          description: List of group members
          items:
            $ref: "#/components/schemas/User"
          minItems: 1
          maxItems: 100
      required:
        - id
        - name
        - members
  responses:
    BadRequest:
      description: The request was not compliant with the documentation
    Unauthorized:
      description: The access token is missing or invalid
    InternalServerError:
      description: The server encountered an internal error